# Conflicts:
#	src/main/java/ru/skypro/homework/controller/AdsController.java
#	src/main/java/ru/skypro/homework/dto/AdsDto.java

+++ src/main/java/ru/skypro/homework/dto/AdsCommentDto.java
@@ -8,14 +8,10 @@ import javax.validation.constraints.Size;
 @Data
 public class AdsCommentDto {
 
-    private int pk;
-
+    private long pk;
     private int author;
-
     private String createdAt;
-
     @NotBlank
     @Size(min = 8)
     private String text;
-
 }
